machines/devices/rasp0/symbol_generator/requirements.txt
pillow==11.3.0
opencv-python==4.12.0.88
numpy==2.2.6